Split runner shell job tree tests
